# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KA.SHA Luxury Fashion eCommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lookbook-outfits:
    get:
      operationId: listLookbookOutfits
      tags: [lookbook]
      summary: List the user's saved outfit combinations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of saved outfits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LookbookOutfit"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createLookbookOutfit
      tags: [lookbook]
      summary: Save a new outfit combination
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLookbookOutfitBody"
      responses:
        "201":
          description: Outfit saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LookbookOutfit"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lookbook-outfits/{id}:
    delete:
      operationId: deleteLookbookOutfit
      tags: [lookbook]
      summary: Delete a saved outfit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Outfit deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lookbook-saved:
    get:
      operationId: listLookbookSaved
      tags: [lookbook]
      summary: List product IDs the user has saved to their lookbook
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of saved product IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: saveLookbookProduct
      tags: [lookbook]
      summary: Save a product to the lookbook (heart it)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveLookbookProductBody"
      responses:
        "201":
          description: Product saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  productId:
                    type: integer
                required:
                  - productId
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lookbook-saved/{productId}:
    delete:
      operationId: unsaveLookbookProduct
      tags: [lookbook]
      summary: Remove a product from the lookbook (un-heart it)
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product removed
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lookbook-photo:
    post:
      operationId: uploadLookbookPhoto
      tags: [lookbook]
      summary: Upload a personal photo to use as the try-on model instead of the AI avatar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
              required: [photo]
      responses:
        "200":
          description: Uploaded photo URL
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url:
                    type: string
        "400":
          description: Invalid upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lookbook-tryon:
    post:
      operationId: submitTryOn
      tags: [lookbook]
      summary: Submit an AI virtual try-on job (top+bottom, or a single dress)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTryOnBody"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TryOnJobAccepted"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lookbook-tryon/{jobId}:
    get:
      operationId: getTryOnJob
      tags: [lookbook]
      summary: Poll the status/result of a try-on job
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TryOnJobStatus"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LookbookLookItem:
      type: object
      required: [productId, role, name, thumbnailUrl]
      properties:
        productId:
          type: integer
        role:
          type: string
          enum: [top, bottom, dress]
        name:
          type: string
        thumbnailUrl:
          type: string
    LookbookOutfit:
      type: object
      required: [id, userId, name, items, gender, resultImageUrl, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/LookbookLookItem"
        gender:
          type: string
          enum: [male, female]
        resultImageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateLookbookOutfitBody:
      type: object
      required: [name, items, gender, resultImageUrl]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        items:
          type: array
          items:
            $ref: "#/components/schemas/LookbookLookItem"
        gender:
          type: string
          enum: [male, female]
        resultImageUrl:
          type: string
    SaveLookbookProductBody:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer
    SubmitTryOnBody:
      type: object
      required: [gender, productIds]
      properties:
        gender:
          type: string
          enum: [male, female]
        productIds:
          type: array
          minItems: 1
          maxItems: 2
          items:
            type: integer
        humanImageUrl:
          description: URL of a user-uploaded photo (from /lookbook-photo) to use instead of the default AI avatar
          type: ["string", "null"]
    TryOnJobAccepted:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, processing, succeeded, failed]
    TryOnJobStatus:
      type: object
      required: [jobId, status, garmentCount, processedCount, resultImageUrl, error]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, processing, succeeded, failed]
        garmentCount:
          type: integer
        processedCount:
          type: integer
        resultImageUrl:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
...